MongoDB: Improve type mapper in schema translator
Now honours:
- DECIMAL => INT64
- INT64 => BIGINT
- TIMESTAMP => TIMESTAMP WITHOUT TIME ZONE

+++ tests/io/mongodb/test_translate.py
@@ -11,7 +11,7 @@ class TestTranslate(unittest.TestCase):
     def test_types_translation(self):
         i = [
             ("DATETIME", "TIMESTAMP WITH TIME ZONE"),
-            ("INT64", "INTEGER"),
+            ("INT64", "BIGINT"),
             ("STRING", "TEXT"),
             ("BOOLEAN", "BOOLEAN"),
             ("INTEGER", "INTEGER"),
